requirements.txt
urllib3~=1.25.8
bs4~=4.9.0
beautifulsoup4~=4.9.0
numpy~=1.18.1
elasticsearch~=7.0.4
bert-serving-client==1.9.8
pandas~=1.0.3
Flask~=1.1.2
